Move functions to OS specific files
